services/cloud/requirements.txt
Django==1.11
Fabric3==1.14.post1
cryptography==2.2.2
pycrypto==2.6.1
sendgrid==5.3.0
django-cors-headers==0.13
django-rest-auth==0.3.0
djangorestframework==3.7
requests==2.4.3
psycopg2==2.8
netifaces==0.10.6
pytz==2013.7
python-dateutil>=2.6.0
